requirements.txt
python-telegram-bot==20.8
Flask==3.0.2
gunicorn==22.0.0
python-dotenv==1.0.1
openai==1.30.1
